Profile, Layouts, Edits.
Changed the Edit form to be more accommodating.

Added a layout.sass file for things like single-column, double-column,
layouts of my own.

Build a profile page. I don’t know if I want it to be a modal or not…

Tomorrow’s the last day to work on this front end mockup, before the
review on Friday.

Then I get to integrate revisions and put in the database code next
week.