Database 1
Integrated a database for ‘proposals’ which takes info from the create
profile form, and publishes it to the gallery and profile.php by id.

Uploading files, editing profiles, and liking comes tomorrow. The day
after I’ll demo these features with the Davenport team.

Then it’s on to integrating the login and register system before the
final launch.

Style still to be done:
-Card transitions on intro sequence
-Steelfish font: heavy and light
-Close out (to everybody.php) on the profile.php background